% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{find_comments}
\alias{find_comments}
\alias{find_comments.data.frame}
\alias{find_comments.character}
\title{Find commented records}
\usage{
find_comments(x, ...)

\method{find_comments}{data.frame}(x, ...)

\method{find_comments}{character}(x, ...)
}
\arguments{
\item{x}{a data frame or character vector}

\item{...}{not used}
}
\value{
A logical vector
}
\description{
Find commented records
}
\examples{
comment <- c(NA, "C", "C", NA, ".", NA, "Comment")
dv <- rnorm(length(comment))
df <- data.frame(C = comment , DV = dv)

find_comments(df)

}
