% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{find_id_col}
\alias{find_id_col}
\title{Find ID column}
\usage{
find_id_col(data)
}
\arguments{
\item{data}{a data.frame to search}
}
\description{
Search data frame names for the first matching candidate ID column name.
See \code{details}.
}
\details{
Column names will be searched against the following candidates
\itemize{
\item \code{getOption("lastdose.id_col")}
\item \code{ID}
\item \code{USUBJID}
\item \code{SUBJID}
\item \code{PTNO}
\item \code{SUBJ}
}

The first the first candidate to be matched will be returned. If there
are no matches, an error is generated.
}
\examples{
data <- data.frame(A = 1, B = 2, PTNO = 3, ID = 4, Z = 99)
lastdose:::find_id_col(data)

}
