% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{find_time_col}
\alias{find_time_col}
\title{Find TIME column}
\usage{
find_time_col(data)
}
\arguments{
\item{data}{a data.frame to search}
}
\description{
Search data frame names for the first matching candidate TIME column name.
See \code{details}.
}
\details{
Column names will be searched against the following candidates
\itemize{
\item \code{TIME}
\item \code{DATETIME}
}

The first the first candidate to be matched will be returned. If there
are no matches, an error is generated.
}
\examples{
data <- data.frame(A = 1, DATETIME = 2, TIME = 3, Z = 99)
lastdose:::find_time_col(data)

}
