% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{lastdose}
\alias{lastdose}
\title{Calculate last dose amount and time after last dose}
\usage{
lastdose(data)
}
\arguments{
\item{data}{data set as data frame}
}
\description{
Calculate last dose amount and time after last dose
}
