% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{find_comments}
\alias{find_comments}
\alias{find_comments.data.frame}
\alias{find_comments.character}
\alias{find_comments.logical}
\title{Find commented records}
\usage{
find_comments(x, ...)

\method{find_comments}{data.frame}(x, ...)

\method{find_comments}{character}(x, ...)

\method{find_comments}{logical}(x, ...)
}
\arguments{
\item{x}{a data frame or character vector}

\item{...}{not used}
}
\value{
A logical vector
}
\description{
This function uses specific criteria for finding comment records that may
not match up with your coding conventions.  See details below as well as
unit tests.
}
\details{
For the data frame method, comments are found in a column with name \code{C}
with type \code{character}.  If a \code{C} column is found that is not character,
a warning is generated and no comments are found.

For the character method or when an appropriate \code{C} column is found when
using the data frame method, a position or row is considered a comment
when it is either \code{NA} or when it is equal to \code{.}.
}
\examples{
comment <- c(NA, "C", "C", NA, ".", NA, "Comment")
dv <- rnorm(length(comment))
df <- data.frame(C = comment , DV = dv, stringsAsFactors = FALSE)

find_comments(df)


}
