% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{lastdose}
\alias{lastdose}
\title{Calculate last dose amount and time after last dose}
\usage{
lastdose(data, fill = -1)
}
\arguments{
\item{data}{data set as data frame}

\item{fill}{time after dose value for records prior to the first dose}
}
\description{
Calculate last dose amount and time after last dose
}
