% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{lastdose}
\alias{lastdose}
\alias{lastdose_list}
\alias{lastdose_df}
\title{Calculate last dose amount and times since previous doses}
\usage{
lastdose(
  data,
  ...,
  include_ldos = TRUE,
  include_tafd = getOption("lastdose.include_tafd", FALSE),
  include_occ = getOption("lastdose.include_occ", TRUE)
)

lastdose_list(
  data,
  time_col = find_time_col(data),
  time_units = getOption("lastdose.time_units", NULL),
  id_col = find_id_col(data),
  fill = -99,
  back_calc = TRUE,
  addl_ties = c("obs_first", "dose_first"),
  comments = find_comments(data),
  include_occ = getOption("lastdose.include_occ", TRUE)
)

lastdose_df(data, ...)
}
\arguments{
\item{data}{data set as data frame; see \code{details}}

\item{...}{arguments passed to \code{\link[=lastdose_list]{lastdose_list()}}}

\item{include_ldos}{\code{logical}; if \code{FALSE} then the \code{LDOS} data is not
appended to the data set.  Only used for the \code{\link[=lastdose]{lastdose()}} function.}

\item{include_tafd}{\code{logical}; if \code{FALSE}, then time after first dose
(\code{TAFD}) data is not appended to the data set; this is only used for the
\code{\link[=lastdose]{lastdose()}} function.}

\item{include_occ}{\code{logical}; if \code{FALSE} then observation occasion counter
(\code{OCC}; see \strong{Details}) is not appended to the data set; this is only
used for the \code{\link[=lastdose]{lastdose()}} function.}

\item{time_col}{character name for the \code{TIME} column; this could be time after
first dose or time after first record or time relative to any origin; input
may be \code{numeric} or \code{POSIXct} (e.g. \code{DATETIME}); if \code{POSIXct}, a numeric
value will be calculated based on the value of \code{time_units}. The data frame
will be searched for the first matching candidate time column using
\code{\link[=find_time_col]{find_time_col()}}; if you don't want \code{lastdose} to search, you should pass
in the name of the column to use for \code{TIME}.}

\item{time_units}{for calculating time when the time column inherits
\code{POSIXct}; you may use any value that is valid for \code{\link[=difftime]{difftime()}}}

\item{id_col}{character name for the subject \code{ID} column; may be numeric
or character; if character, a numeric value is derived. The data frame
will be searched for the first matching candidate \code{ID} column using
\code{\link[=find_id_col]{find_id_col()}}; if you don't want \code{lastdose} to search, you should pass
in the name of the column to use for \code{ID}.}

\item{fill}{the value for \code{TAD} and \code{TAFD} that is used for records when no
doses are found for an individual or when \code{back_calc} is \code{FALSE}.}

\item{back_calc}{if \code{TRUE}, then the time before the first dose
is calculated for records prior to the first dosing record when
at least one dosing record is found in the data set.  Records before
the first dosing record will have negative values.}

\item{addl_ties}{what to do when doses scheduled through \code{ADDL} happen at
the same time as observation records; if \code{obs_first} then the observation
is assumed to happen before the dose and the observation is a trough
concentration; if \code{dose_first} then the dose is assumed to be administered
and the observation made immediately after (with no advance in time). See
details.}

\item{comments}{a logical vector with length equal to the number of rows
in \code{data} indicating which records are to be ignored when looking for \code{TAD}
and \code{LDOS}.}
}
\description{
This function calculates the last dose amount (\code{LDOS}), the time after
last dose (\code{TAD}), time after first dose (\code{TAFD}), and observation
occasion (\code{OCC}). Use \code{\link[=lastdose]{lastdose()}}
to add (or potentially replace) columns to the input data frame;
\code{\link[=lastdose_list]{lastdose_list()}} and \code{\link[=lastdose_df]{lastdose_df()}} returns calculated information
as either \code{list} or \code{data.frame} format without modifying the input data.
}
\details{
When calling \code{\link[=lastdose]{lastdose()}} to modify the data frame, two columns will be
added (by default): \code{TAD} indicating the time after the most-recent dose,
and \code{LDOS} indicating the amount of the most recent dose. \code{TAFD} indicating
the time after the first dose record (\code{EVID} 1 or 4) can be added via the
\code{include_tafd} argument and users can opt out from adding \code{LDOS} with the
\code{include_ldos} argument.

When calling \code{\link[=lastdose_list]{lastdose_list()}} or \code{\link[=lastdose_df]{lastdose_df()}}, the respective items are
accessible with \code{tad},  \code{tafd}, and \code{ldos} (note the lower case form here to
distinguish from the columns that might be added to the data frame).

\strong{Time after first dose (TAFD)}: note that time after first dose (\code{TAFD})
is the time after the first dosing record (\code{EVID} 1 or 4) in the data frame
that you pass in. If you don't have a dosing record for the first dose to
anchor this calculation, you should opt out.

\strong{Occasion (OCC)}: observation occasions (\code{OCC}) occur when there is an
observation record (with \code{EVID=0}) following a dose record (\verb{EVID 1 or 4});
\code{OCC} starts at \code{0} and increments with each dose that is followed by at
least one observation record. The \code{OCC} calculation ignores all commented
records (doses or observations).

\strong{Handling of commented records}: Dosing records that have been "commented"
(as indicated with the \code{comments} argument) will never be considered as
actual doses when determining \code{TAD}, \code{TAFD}, and \code{LDOS}.  But commented
records (doses and non-doses) will be assigned \code{TAD}, \code{TAFD}, and \code{LDOS}
according to the last non-commented dosing record.

\strong{Additional notes}:
\itemize{
\item All functions require an input data set as a data frame
\item The data set should be formatted according to \code{NMTRAN} type
conventions
\item Required columns
\itemize{
\item A subject ID column (either \code{ID} or user-specified)
\item A record time column (either \code{TIME} or user-specified)
\item \code{AMT} or \code{amt}: dose amount for dosing records
\item \code{EVID} or \code{evid}: event ID; records with \code{EVID} or 1 or 4
are considered dosing records
}
\item Optional columns
\itemize{
\item \code{ADDL} or \code{addl}: additional doses to administer
\item \code{II} or \code{ii}: dosing interval
}
\item An error is generated if required columns are not found; no error
or warning if optional columns are not found
\item All required and optional columns are required to be numeric
\item Missing values are not allowed in: \code{ID}, \code{EVID}, \code{ADDL}, \code{II}
\item When missing values are found in \code{TIME}, both \code{TAD} and \code{LDOS} are set to
missing
\item An error is generated for missing \code{AMT} in dosing records (evid 1 or 4)
\item No error is generated for missing \code{AMT} in non-dosing records
}

An example illustrating the \code{addl_ties} argument: when there is \code{Q24h}
dosing and both an an additional dose and an observation happen at 24 hours,
\code{obs_first} will set the observation \code{TAD} to 24 and \code{dose_first} will set
the observation \code{TAD} to 0.
}
\section{Options}{


These are options that can be set to customize \code{lastdose} behavior
for the current context.  See \code{?options} for how to set an option.
\itemize{
\item \code{lastdose.time_units}: sets the default time unit that is used to calculate
relative times when the time column is represented as date-time data
(\code{POSIXct})
\item \code{lastdose.id_col}: sets the default value for the \code{id_col} argument
to last dose; this identifies the column that is to be used to distinguish
individuals; the data in this column may be numeric or character
\item \code{lastdose.include_tafd}: sets default value for \code{include_tafd}; if \code{TRUE}
then the time since the first dose record (EVID 1 or EVID 4) in the data
set will be automatically appended to the output data frame when
calling \code{lastdose()}; \code{tafd} is always included when calling
\code{lastdose_df()} and \code{lastdose_list()}
}
}

\examples{
file <- system.file("csv/data1.csv", package="lastdose")

require("Rcpp")

data <- read.csv(file)

a <- lastdose(data)

b <- lastdose_df(data)

c <- lastdose_list(data)


}
