% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{lastdose}
\alias{lastdose}
\alias{lastdose_list}
\alias{lastdose_df}
\title{Calculate last dose amount (LDOS) and time after last dose (TAD)}
\usage{
lastdose(data, ..., include_ldos = TRUE)

lastdose_list(
  data,
  fill = -99,
  back_calc = TRUE,
  addl_ties = c("obs_first", "dose_first"),
  comments = find_comments(data),
  fill_comments_na = FALSE
)

lastdose_df(data, ...)
}
\arguments{
\item{data}{data set as data frame; see \code{details}}

\item{...}{arguments passed to \link{lastdose_list}}

\item{include_ldos}{\code{logical}; if \code{FALSE} then the \code{LDOS} data is not
appended to the data set.  Only used for the \link{lastdose} function.}

\item{fill}{the value for \code{TAD} that is used for records when no
doses are found for an individual or when \code{back_calc} is \code{FALSE}.}

\item{back_calc}{if \code{TRUE}, then the time before the first dose
is calculated for records prior to the first dosing record when
at least one dosing record is found in the data set.  Records before
the first dosing record will have negative values.}

\item{addl_ties}{what to do when doses scheduled through \code{ADDL} happen at
the same time as observation records; if \code{obs_first} then the observation
is assumed to happen before the dose and the observation is a trough
concentration; if \code{dose_first} then the dose is assumed to be administered
and the observation made immediately after (with no advance in time). See
details.}

\item{comments}{a logcial vector with length equal to the number of rows
in \code{data} indicating which records are to be ignored when looking for \code{TAD}
and \code{LDOS}.  See all the \code{fill_comments_na} argument.}

\item{fill_comments_na}{if \code{TRUE}, then \code{TAD} and \code{LDOS} will be filled with
\code{NA} when the comment indicator is \code{TRUE}; otherwise, commented records
are filled in with the time since and amount of the dose on the last
non-commented dosing record (the default).}
}
\description{
Use \link{lastdose} to add (or potentially replace) columns to the input
data frame; \link{lastdose_list} and \link{lastdose_df} returns calculated information
as either \code{list} or \code{data.frame} format without modifying the input data.
}
\details{
When calling \link{lastdose} to modify the data frame, two columns will be
added (by default): \code{TAD} indicating the time after the most-recent dose
and \code{LDOS} indicating the amount of the most recent dose.  This default
behavior can be modified with the \code{include_ldos} argument.

When calling \link{lastdose_list} or \link{lastdose_df}, the respective items are
accessible with \code{tad} and \code{ldos} (note the lower case form here to
distinguish from the columns that might be added to the data frame).

\strong{Handling of commented records}: Dosing records that have been "commented"
(as indicated with the \code{comments} argument) will never be considered as
actual doses when determining \code{TAD} and \code{LDOS}.  But commented records (doses
and non-doses) will be assigned \code{TAD} and \code{LDOS} according to the last
non-commented dosing record by default.  The behavior can be modified by
setting through the \code{fill_comments_na} argument.

\strong{Additional notes}:
\itemize{
\item All functions require an input data set as a data frame
\item The data set should be formatted according to \code{NMTRAN} type
conventions
\item Required columns
\itemize{
\item \code{ID} or \code{id}: subject identifier
\item \code{TIME} or \code{time}: data set time
\item \code{AMT} or \code{amt}: dose amount for dosing records
\item \code{EVID} or \code{evid}: event ID; records with \code{EVID} or 1 or 4
are considered dosing records
}
\item Optional columns
\itemize{
\item \code{ADDL} or \code{addl}: additional doses to administer
\item \code{II} or \code{ii}: dosing interval
}
\item An error is generated if required columns are not found; no error
or warning if optional columns are not found
\item All required and optional columns are required to be numeric
\item Missing values are not allowed in: \code{ID},\code{EVID},\code{ADDL},\code{II}
\item When missing values are found in \code{TIME}, both \code{TAD} and \code{LDOS} are set to
missing
\item An error is generated for missing \code{AMT} in dosing records (evid 1 or 4)
\item No error is generated for missing \code{AMT} in non-dosing records
}

An example illustrating the \code{addl_ties} argument: when there is \code{Q24h}
dosing and both an an additional dose and an observation happen at 24 hours,
\code{obs_first} will set the observation\code{TAD} to 24 and \code{dose_first} will set
the observation \code{TAD} to 0.
}
\examples{
file <- system.file("csv/data1.csv", package="lastdose")

require("Rcpp")

data <- read.csv(file)

a <- lastdose(data)

b <- lastdose_df(data)

c <- lastdose_list(data)


}
