% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastdose.R
\name{lastdose}
\alias{lastdose}
\alias{lastdose_list}
\alias{lastdose_df}
\title{Calculate last dose amount and times since previous doses}
\usage{
lastdose(data, ..., include_ldos = TRUE, include_tafd = TRUE)

lastdose_list(
  data,
  time_col = "TIME",
  time_units = getOption("lastdose.time_units", NULL),
  id_col = getOption("lastdose.id_col", "ID"),
  fill = -99,
  back_calc = TRUE,
  addl_ties = c("obs_first", "dose_first"),
  comments = find_comments(data)
)

lastdose_df(data, ...)
}
\arguments{
\item{data}{data set as data frame; see \code{details}}

\item{...}{arguments passed to \code{\link[=lastdose_list]{lastdose_list()}}}

\item{include_ldos}{\code{logical}; if \code{FALSE} then the \code{LDOS} data is not
appended to the data set.  Only used for the \code{\link[=lastdose]{lastdose()}} function.}

\item{include_tafd}{\code{logical}; if \code{FALSE}, then \code{TAFD} data is not appended
to the data set.  Only used for the \code{\link[=lastdose]{lastdose()}} function.}

\item{time_col}{character name for the \code{TIME} column; this could be time after
first dose or time after first record or time relative to any origin; input
may be \code{numeric} or \code{POSIXct} (e.g. \code{DATETIME}); if \code{POSIXct}, a numeric
value will be calculated based on the value of \code{time_units}}

\item{time_units}{for calculating time when the time column inherits
\code{POSIXct}; you may use any value that is valid for \code{\link[=difftime]{difftime()}}}

\item{id_col}{character name for the subject \code{ID} column; may be numeric
or character; if character, a numeric value is derived}

\item{fill}{the value for \code{TAD} and \code{TAFD} that is used for records when no
doses are found for an individual or when \code{back_calc} is \code{FALSE}.}

\item{back_calc}{if \code{TRUE}, then the time before the first dose
is calculated for records prior to the first dosing record when
at least one dosing record is found in the data set.  Records before
the first dosing record will have negative values.}

\item{addl_ties}{what to do when doses scheduled through \code{ADDL} happen at
the same time as observation records; if \code{obs_first} then the observation
is assumed to happen before the dose and the observation is a trough
concentration; if \code{dose_first} then the dose is assumed to be administered
and the observation made immediately after (with no advance in time). See
details.}

\item{comments}{a logical vector with length equal to the number of rows
in \code{data} indicating which records are to be ignored when looking for \code{TAD}
and \code{LDOS}.  See all the \code{fill_comments_na} argument.}
}
\description{
This function calculates the last dose amount (\code{LDOS}), the time after
last dose (\code{TAD}), and time after first dose (\code{TAFD}). Use \code{\link[=lastdose]{lastdose()}}
to add (or potentially replace) columns to the input data frame;
\code{\link[=lastdose_list]{lastdose_list()}} and \code{\link[=lastdose_df]{lastdose_df()}} returns calculated information
as either \code{list} or \code{data.frame} format without modifying the input data.
}
\details{
When calling \code{\link[=lastdose]{lastdose()}} to modify the data frame, three columns will be
added (by default): \code{TAD} indicating the time after the most-recent dose,
\code{TAFD} indicating the time after the first dose and \code{LDOS} indicating the
amount of the most recent dose.  This default behavior can be modified with
the \code{include_ldos} and \code{include_tafd} arguments.

When calling \code{\link[=lastdose_list]{lastdose_list()}} or \code{\link[=lastdose_df]{lastdose_df()}}, the respective items are
accessible with \code{tad},  \code{tafd}, and \code{ldos} (note the lower case form here to
distinguish from the columns that might be added to the data frame).

\strong{Handling of commented records}: Dosing records that have been "commented"
(as indicated with the \code{comments} argument) will never be considered as
actual doses when determining \code{TAD}, \code{TAFD}, and \code{LDOS}.  But commented
records (doses and non-doses) will be assigned \code{TAD}, \code{TAFD}, and \code{LDOS}
according to the last non-commented dosing record.

\strong{Additional notes}:
\itemize{
\item All functions require an input data set as a data frame
\item The data set should be formatted according to \code{NMTRAN} type
conventions
\item Required columns
\itemize{
\item A subject ID column (either \code{ID} or user-specified)
\item A record time column (either \code{TIME} or user-specified)
\item \code{AMT} or \code{amt}: dose amount for dosing records
\item \code{EVID} or \code{evid}: event ID; records with \code{EVID} or 1 or 4
are considered dosing records
}
\item Optional columns
\itemize{
\item \code{ADDL} or \code{addl}: additional doses to administer
\item \code{II} or \code{ii}: dosing interval
}
\item An error is generated if required columns are not found; no error
or warning if optional columns are not found
\item All required and optional columns are required to be numeric
\item Missing values are not allowed in: \code{ID}, \code{EVID}, \code{ADDL}, \code{II}
\item When missing values are found in \code{TIME}, both \code{TAD} and \code{LDOS} are set to
missing
\item An error is generated for missing \code{AMT} in dosing records (evid 1 or 4)
\item No error is generated for missing \code{AMT} in non-dosing records
}

An example illustrating the \code{addl_ties} argument: when there is \code{Q24h}
dosing and both an an additional dose and an observation happen at 24 hours,
\code{obs_first} will set the observation \code{TAD} to 24 and \code{dose_first} will set
the observation \code{TAD} to 0.
}
\section{Options}{


These are options that can be set to customize \code{lastdose} behavior
for the current context.  See \code{?options} for how to set an option.
\itemize{
\item \code{lastdose.time_units}: sets the default time unit that is used to calculate
relative times when the time column is represented as date-time data
(\code{POSIXct})
\item \code{lastdose.id_col}: sets the default value for the \code{id_col} argument
to last dose; this identifies the column that is to be used to distinguish
individuals; the data in this column may be numeric or character
}
}

\examples{
file <- system.file("csv/data1.csv", package="lastdose")

require("Rcpp")

data <- read.csv(file)

a <- lastdose(data)

b <- lastdose_df(data)

c <- lastdose_list(data)


}
